+++ packages/terra-button/tests/wdio/button-spec.js
@@ -255,4 +255,12 @@ Terra.describeViewports('Button', ['tiny'], () => {
       Terra.validates.element('programmatic re-enabled -Enabled Button is not focused', { selector: '#root' });
     });
   });
+
+  describe('Selectable button', () => {
+    it('displays selected button', () => {
+      browser.url('/raw/tests/cerner-terra-core-docs/button/selectable-button');
+      $('#button1').click();
+      Terra.validates.element('selected-button');
+    });
+  });
 });