+++ packages/terra-dropdown-button/tests/jest/SplitButton.test.jsx
@@ -3,10 +3,21 @@ import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 import { IntlProvider } from 'react-intl';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
 import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { v4 as uuidv4 } from 'uuid';
 import translationsFile from '../../translations/en.json';
 
 /* eslint-disable-next-line import/no-extraneous-dependencies */
 import SplitButton, { Item } from '../../src/SplitButton';
+
+let mockSpyUuid;
+beforeAll(() => {
+  mockSpyUuid = jest.spyOn(uuidv4, 'v4').mockReturnValue('00000000-0000-0000-0000-000000000000');
+});
+
+afterAll(() => {
+  mockSpyUuid.mockRestore();
+});
 
 describe('Dropdown Button', () => {
   it('should render a default split type', () => {