+++ packages/terra-form-select/src/SearchSelect.jsx
@@ -142,6 +142,7 @@ class SearchSelect extends React.Component {
     this.handleChange = this.handleChange.bind(this);
     this.handleDeselect = this.handleDeselect.bind(this);
     this.handleSelect = this.handleSelect.bind(this);
+    this.handleResetComboboxValue = this.handleResetComboboxValue.bind(this);
   }
 
   /**
@@ -184,6 +185,13 @@ class SearchSelect extends React.Component {
   }
 
   /**
+   * Added to reset value field for combobox
+   */
+  handleResetComboboxValue() {
+    this.setState({ value: '' });
+  }
+
+  /**
    * Returns the appropriate variant display
    */
   display() {
@@ -222,6 +230,7 @@ class SearchSelect extends React.Component {
         totalOptions={SelectUtil.getTotalNumberOfOptions(children)}
         clearOptionDisplay={clearOptionDisplay}
         inputId={inputId}
+        resetComboboxValue={this.handleResetComboboxValue}
       >
         {children}
       </Frame>