+++ packages/terra-action-header/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 2.87.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 2.86.0 - (November 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 2.85.0 - (October 23, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.84.0 - (October 3, 2023)
 
 * Changed