+++ packages/terra-alert/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 4.86.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.85.0 - (November 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.84.0 - (October 23, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.83.0 - (October 16, 2023)
 
 * Updated