+++ packages/terra-form-fieldset/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 2.78.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.77.0 - (October 16, 2023)
 
 * Changed