+++ packages/terra-list/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 4.67.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.66.0 - (October 23, 2023)
+
+* Added
+  * Added `isTabFocusDisabled` prop to disable tab key navigation.
+
 * Fixed
   * Fixed _property is undefined_ error while navigating with a keyboard.