+++ packages/terra-section-header/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 2.64.0 - (November 13, 2023)
+
+  * Changed
+    * Updated the section header components to meet accessibility standards.
+
+## 2.63.0 - (October 23, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.62.0 - (August 23, 2023)
 
 * Changed