+++ packages/terra-status-view/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 4.69.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.68.0 - (November 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.67.0 - (October 23, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.66.0 - (October 16, 2023)
 
 * Changed