+++ packages/terra-status/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 4.62.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.61.0 - (October 23, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.60.0 - (October 16, 2023)
 
 * Changed