+++ packages/terra-toggle-button/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 3.81.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.80.0 - (November 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.79.0 - (October 23, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.78.0 - (October 16, 2023)
 
 * Changed