+++ packages/terra-toolbar/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 1.41.0 - (November 13, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.40.0 - (November 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.39.0 - (October 23, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.38.0 - (October 3, 2023)
 
 * Changed